pyproject.toml
[tool.poetry]
name = "djangosrlmongo"
version = "0.1.0"
description = ""
authors = ["shaheerzaman <shaheer_zaman84@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.2"
Django = "3.0.5"
djongo = "^1.3.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
